Seção 17: Palestrantes e Redes Sociais - Frontend

+++ Back/src/ProEventos.API/Helpers/Util.cs
@@ -13,8 +13,7 @@ namespace ProEventos.API.Helpers
         private readonly IWebHostEnvironment _hostEnvironment;
         public Util(IWebHostEnvironment hostEnvironment)
         {
-            _hostEnvironment = hostEnvironment;
-            
+            _hostEnvironment = hostEnvironment; 
         }
 
         public async Task<string> SaveImage(IFormFile imageFile, string destino)
@@ -26,7 +25,7 @@ namespace ProEventos.API.Helpers
 
             imageName = $"{imageName}{DateTime.UtcNow.ToString("yymmssfff")}{Path.GetExtension(imageFile.FileName)}";
 
-            var imagePath = Path.Combine(_hostEnvironment.ContentRootPath, @"Resources/{destino}", imageName);
+            var imagePath = Path.Combine(_hostEnvironment.ContentRootPath, @$"Resources/{destino}", imageName);
 
             using (var fileStream = new FileStream(imagePath, FileMode.Create))
             {
@@ -38,9 +37,12 @@ namespace ProEventos.API.Helpers
 
         public void DeleteImage(string imageName, string destino)
         {
-            var imagePath = Path.Combine(_hostEnvironment.ContentRootPath, @"Resources/{destino}", imageName);
-            if (System.IO.File.Exists(imagePath))
-                System.IO.File.Delete(imagePath);
+            if (!string.IsNullOrEmpty(imageName))
+            {
+                var imagePath = Path.Combine(_hostEnvironment.ContentRootPath, @$"Resources/{destino}", imageName);
+                if (System.IO.File.Exists(imagePath))
+                    System.IO.File.Delete(imagePath);
+            }
         }
     }
 }
